% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp_modelling.R
\name{ACCIRFconf}
\alias{ACCIRFconf}
\title{Multivariate normal random series}
\usage{
ACCIRFconf(IRF)
}
\arguments{
\item{IRF}{is an impulse response function}
}
\value{
an accumulative impulse response function with the same dimension of IRF.
}
\description{
This function calculates the accumulative impulse response function.
}
