% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp_modelling.R
\name{BOOT}
\alias{BOOT}
\title{BOOT}
\usage{
BOOT(C, nrun)
}
\arguments{
\item{C}{list of resid used as container for bootstrapped residuals}

\item{nrun}{number of bootstrap runs.}
}
\value{
an array to contain the list of bootstrap and number of bootstrap runs.
}
\description{
This creates a list of PRP object to contain the bootstrap results.
}
