% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp_modelling.R
\name{bstrp}
\alias{bstrp}
\title{bstrp}
\usage{
bstrp(Model, nrun, Method, T)
}
\arguments{
\item{Model}{an estimated Model object.}

\item{nrun}{number of bootstrap runs.}

\item{Method}{a parameter that determines how the bootstrap residuals are to to generate. Method="residuals" for residuals' bootstrap, Method="norm" for parametric bootstrap from a normal distribution, and Method = "binomial" for parametric bootstrap from a binomial distribution.}

\item{T}{is the length of the bootstrap residuals.}
}
\value{
a list of Model residuals.
}
\description{
This function calculates bootstrap residuals.
}
