% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{dgp}
\alias{dgp}
\title{Example of a dgp function}
\usage{
dgp(Model, T, M = 1)
}
\arguments{
\item{Model}{is a Model object with given paramters (PARAM) and resid or Model$RD="norm"}

\item{T}{the number of data points to be generated.}

\item{M}{usually takes the value 1.}
}
\value{
A list of Model objact with generated data (ENDOG) and the parameters used to generate the data.
}
\description{
This function shows how a dgp can be constructed.
}
