% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{est}
\alias{est}
\title{Example of a est function}
\usage{
est(Model)
}
\arguments{
\item{Model}{is a Model object with generated data.}
}
\value{
A list of Model objact with estimated parameter (PARAM) values and the generated data (ENDOG).
}
\description{
This function shows how an est can be constructed to estimate the model parameter in PARAM
}
