% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp_modelling.R
\name{MODEL.BOOT}
\alias{MODEL.BOOT}
\title{MODEL.BOOT}
\usage{
MODEL.BOOT(Model, PRP, PRP.PARAM, nrun, Method)
}
\arguments{
\item{Model}{object of a complete Model with known RD. RD="resid" non-parametric bootstrap, RD="norm" parametric boostrap.}

\item{PRP}{object of the MODEL.PRO output which will be boostraped.}

\item{PRP.PARAM}{the parameter that is used to calculate the PRP.}

\item{nrun}{number of bootrap runs}

\item{Method}{a parameter that determines how the boostrap samples are generated.}
}
\value{
a list of bootstraped PARAM, bootstraped resid and bootstraped PRP.
}
\description{
This function bootstraps the Properties of a complete Model inlcuding bootstrap of the estimated parametrs. Often it is used for model validation purpose.
}
