% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp_modelling.R
\name{MODEL.EST}
\alias{MODEL.EST}
\title{MODEL.EST}
\usage{
MODEL.EST(Model)
}
\arguments{
\item{Model}{object of an incomplete Model with known H_PARAM, ENDOG and EXOG.}
}
\value{
Model object with estimated PARAM, INFOC, and resid.
}
\description{
This function completes the Model by providing PARAM and resid for given endogeneous and exogeneous variables.
In application it will be implemented through est(Model).
}
