% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp_modelling.R
\name{MODEL.PRP}
\alias{MODEL.PRP}
\title{MODEL.PRP}
\usage{
MODEL.PRP(Model, PRP.PARAM)
}
\arguments{
\item{Model}{object of a complete Model with known H_PARAM, PARAM, ENDOG, EXOG and resid.}

\item{PRP.PARAM}{a list containing parameters that must be used to calculate the properties.}
}
\value{
a list of PRP of interest.
}
\description{
This function calculates certain properties of a complete Model for given PRP.PARAM. The properties are usually for some application purposes.
In application it will be implemented through prp(Model,PRP.PARAM).
}
