% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp_modelling.R
\name{MODEL.SELECT}
\alias{MODEL.SELECT}
\title{MODEL.SELECT}
\usage{
MODEL.SELECT(Model, H_PARAML, K)
}
\arguments{
\item{Model}{object of a incomplete Model with a set of K H_PARAMs, ENDOG and EXOG.}

\item{H_PARAML}{a list of K H_PARAMs.}

\item{K}{number of H_PARAM for selection.}
}
\value{
a array of K model selection values.
}
\description{
This function calculates model selection values for a set of given H_PARAMs for selection of the hypo-parameter.
}
