% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{prp}
\alias{prp}
\title{Example of a prp function}
\usage{
prp(Model, PRP.PARAM)
}
\arguments{
\item{Model}{is an estimated Model object.}

\item{PRP.PARAM}{a list containing the parameters of the to be tested properties. Typitcally it includs the values of the test statistics under the null hypothesis (Cshareo).}
}
\value{
A list of bootstraped test statistics.
}
\description{
This function shows how a prp function can be constructed to test the designed properties.
}
