% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp_modelling.R
\name{Summary}
\alias{Summary}
\title{Summary of bootstrap results}
\usage{
Summary(OUT = bootresult, Model = EST, PRP.PARAM)
}
\arguments{
\item{OUT}{is the output of MODEL.BOOT}

\item{Model}{is an estimated Model object}

\item{PRP.PARAM}{is the bootstrap parameter used in MODEL.BOOT}
}
\value{
a list containing parameter bootstrapped confidence interval, the bootstrapped interval of test statistics in Cshare, and the impulse response functions.
}
\description{
This function summarizes the bootstrap results.
}
